refactor(#zimic): move browser and ndoe interceptors to dedicated directories
