+++ packages/zimic/src/interceptor/http/HttpInterceptor/browser/BrowserHttpInterceptor.ts
@@ -1,7 +1,7 @@
-import BrowserHttpInterceptorWorker from '../HttpInterceptorWorker/BrowserHttpInterceptorWorker';
-import HttpInterceptor from './HttpInterceptor';
-import { HttpInterceptorOptions } from './types/options';
-import { HttpInterceptorSchema } from './types/schema';
+import BrowserHttpInterceptorWorker from '../../HttpInterceptorWorker/BrowserHttpInterceptorWorker';
+import HttpInterceptor from '../HttpInterceptor';
+import { HttpInterceptorOptions } from '../types/options';
+import { HttpInterceptorSchema } from '../types/schema';
 
 class BrowserHttpInterceptor<Schema extends HttpInterceptorSchema> extends HttpInterceptor<Schema> {
   constructor(options: HttpInterceptorOptions) {
