+++ packages/zimic/src/interceptor/http/HttpInterceptor/browser/InternalBrowserHttpInterceptor.ts
@@ -1,5 +1,5 @@
+import { HttpInterceptorSchema } from '../types/schema';
 import BrowserHttpInterceptor from './BrowserHttpInterceptor';
-import { HttpInterceptorSchema } from './types/schema';
 
 class InternalBrowserHttpInterceptor<Schema extends HttpInterceptorSchema> extends BrowserHttpInterceptor<Schema> {
   worker() {
