+++ packages/zimic/src/interceptor/http/HttpInterceptor/node/InternalNodeHttpInterceptor.ts
@@ -1,5 +1,5 @@
+import { HttpInterceptorSchema } from '../types/schema';
 import NodeHttpInterceptor from './NodeHttpInterceptor';
-import { HttpInterceptorSchema } from './types/schema';
 
 class InternalNodeHttpInterceptor<Schema extends HttpInterceptorSchema> extends NodeHttpInterceptor<Schema> {
   worker() {
