+++ packages/zimic/src/interceptor/http/HttpInterceptor/node/NodeHttpInterceptor.ts
@@ -1,7 +1,7 @@
-import NodeHttpInterceptorWorker from '../HttpInterceptorWorker/NodeHttpInterceptorWorker';
-import HttpInterceptor from './HttpInterceptor';
-import { HttpInterceptorOptions } from './types/options';
-import { HttpInterceptorSchema } from './types/schema';
+import NodeHttpInterceptorWorker from '../../HttpInterceptorWorker/NodeHttpInterceptorWorker';
+import HttpInterceptor from '../HttpInterceptor';
+import { HttpInterceptorOptions } from '../types/options';
+import { HttpInterceptorSchema } from '../types/schema';
 
 class NodeHttpInterceptor<Schema extends HttpInterceptorSchema> extends HttpInterceptor<Schema> {
   constructor(options: HttpInterceptorOptions) {
