+++ packages/zimic/src/interceptor/http/HttpInterceptor/types/classes.ts
@@ -1,4 +1,4 @@
-import BrowserHttpInterceptor from '../BrowserHttpInterceptor';
-import NodeHttpInterceptor from '../NodeHttpInterceptor';
+import BrowserHttpInterceptor from '../browser/BrowserHttpInterceptor';
+import NodeHttpInterceptor from '../node/NodeHttpInterceptor';
 
 export type HttpInterceptorClass = typeof BrowserHttpInterceptor | typeof NodeHttpInterceptor;
